Rename filters/parser/* to filters/lang/*
