+++ src/controllers/filters/lang/FilterParser.cpp
@@ -2,13 +2,13 @@
 
 #include "Application.hpp"
 #include "common/Channel.hpp"
-#include "controllers/filters/parser/expressions/BinaryOperation.hpp"
-#include "controllers/filters/parser/expressions/Expression.hpp"
-#include "controllers/filters/parser/expressions/ListExpression.hpp"
-#include "controllers/filters/parser/expressions/RegexExpression.hpp"
-#include "controllers/filters/parser/expressions/UnaryOperation.hpp"
-#include "controllers/filters/parser/expressions/ValueExpression.hpp"
-#include "controllers/filters/parser/Types.hpp"
+#include "controllers/filters/lang/expressions/BinaryOperation.hpp"
+#include "controllers/filters/lang/expressions/Expression.hpp"
+#include "controllers/filters/lang/expressions/ListExpression.hpp"
+#include "controllers/filters/lang/expressions/RegexExpression.hpp"
+#include "controllers/filters/lang/expressions/UnaryOperation.hpp"
+#include "controllers/filters/lang/expressions/ValueExpression.hpp"
+#include "controllers/filters/lang/Types.hpp"
 #include "messages/Message.hpp"
 #include "providers/twitch/TwitchBadge.hpp"
 #include "providers/twitch/TwitchChannel.hpp"
