+++ src/controllers/filters/lang/FilterParser.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
-#include "controllers/filters/parser/expressions/Expression.hpp"
-#include "controllers/filters/parser/Tokenizer.hpp"
-#include "controllers/filters/parser/Types.hpp"
+#include "controllers/filters/lang/expressions/Expression.hpp"
+#include "controllers/filters/lang/Tokenizer.hpp"
+#include "controllers/filters/lang/Types.hpp"
 
 namespace chatterino {
