+++ src/controllers/filters/lang/Tokenizer.cpp
@@ -1,4 +1,4 @@
-#include "controllers/filters/parser/Tokenizer.hpp"
+#include "controllers/filters/lang/Tokenizer.hpp"
 
 #include "common/QLogging.hpp"
