+++ src/controllers/filters/lang/Types.cpp
@@ -1,7 +1,7 @@
 #include "Types.hpp"
 
-#include "controllers/filters/parser/expressions/Expression.hpp"
-#include "controllers/filters/parser/Tokenizer.hpp"
+#include "controllers/filters/lang/expressions/Expression.hpp"
+#include "controllers/filters/lang/Tokenizer.hpp"
 
 namespace chatterino::filters {
