+++ src/controllers/filters/lang/expressions/BinaryOperation.hpp
@@ -1,14 +1,14 @@
 #pragma once
 
-#include "controllers/filters/parser/expressions/Expression.hpp"
-#include "controllers/filters/parser/Types.hpp"
+#include "controllers/filters/lang/expressions/Expression.hpp"
+#include "controllers/filters/lang/Types.hpp"
 
 namespace chatterino::filters {
 
-class UnaryOperation : public Expression
+class BinaryOperation : public Expression
 {
 public:
-    UnaryOperation(TokenType op, ExpressionPtr right);
+    BinaryOperation(TokenType op, ExpressionPtr left, ExpressionPtr right);
 
     QVariant execute(const ContextMap &context) const override;
     PossibleType synthesizeType() const override;
@@ -17,6 +17,7 @@ public:
 
 private:
     TokenType op_;
+    ExpressionPtr left_;
     ExpressionPtr right_;
 };
